Multiple Analogs + Code Refinements
Added the analog structure based on the looms
All 8 inputs (for now only 6 activated) are  on ADC1 running at 71.5 cycles for each conversion and putting it all in a big buffer.
Averaging function modified to accommodate for the change in the ADC buffer topology.
mapped some analogs to the inputs.
canceled emergency button analog functionality that will not be implemented in the end.
 Defined the 12V internal voltage measurement gain and assigned SHIFTERAnalog06 as the input
Implemented the check for the VUpDnButton from the SIU and defined the upper and lower limits for each actuation.